Begin adding git for changeset tracking
This required a significant refactor/rethink of the organization
and responsibilities of the different classes, especially as I
have needed to reconsider exactly what the use cases are and how
they are achieved. Specifically, the workspace concept and type
have been added as a sandbox location for authoring and managing
environment changes from a project-type directory, rather than
all global data and data storage in the cli.