Remove git integration and improve workflow
Another major change, the git integration was not working and seemed
couter to the overall goals of environment and package management,
leaning more towards development but that can be separately managed.

Also replaces the fuse-overlayfs usage with a set of provileged
binaries that use the kernels overlay implementation and a mount
namespace to isolate the process. This allows everything to be
consistently mounted at /env, making it accessible and visible.