Add low level package to manage object graph
This package defines a set of interfaces and supporting methods for
digraph definition, traversal etc. This will serve as the basis for
definiting all the api objects in spfs so that clean-type operations can
be pure digraph operations and not need to type switch of know about
the implementation deatils of the objects themselves.