Allow remote addresses to be given at cmd line
This updates the config.get_remote function so that is can load remote
addresses directly. Although they are not from the config, the config is
the central decision maker for loading remote repositories. This
enables the easy export and import of tag data in CI by simply pushing
and pulling from a local repo withing the dist folder