Bring in tracking/graph/storage types
I needed to move the concrete storage types in to graph in order to make
the interface reasonable and provide an enum rather than a trait. Since
there is no real upcasting in rust this is the better way. It also resolves
some annoyingness that was in the original API, anyway and feels like a
more natural place for these data types