Example package spec for LLVM+Clang 12
There are a few oddities here that I'll just admit and hope for
forgiveness:

1. The build instructions are complex enough that I have a separate
build.sh instead of describing the build steps explicitly in the
spk.yaml file. (A lot of that complexity is ensuring that LLVM is
built with all the options that are needed downstream by OSL, and
in particular, OSL with Cuda/OptiX support. Much of that doesn't
necessarily happen with an LLVM build with only default choices.)

2. The build.sh file is carefully constructed to be a valid build
recipe for both spk and Rez. (That's how I needed it to be at SPI
rather than have two similar scripts that could accidentally
diverge. But eventually when we are spk only, I'll remove the extra
complexity.)

3. In the build.sh, there is one bit of SPI-only magic, where I prefer
to build llvm with a specific older version of clang, at a known file
path location. You can safely ignore that, I believe I have correctly
written the script so that if that is not found, it'll safely fall
back to building llvm+clang with the active gcc.

I'm leaving these ugly bits in, partly out of laziness, but also
partly because it's worth having people see how these things look in
the wild (and maybe when written by a beginner just trying to cobble
it all together to get work done), and for the sake of being able (for
now, at least) to preserve my ability to keep this example and my
actual internal build scripts in sync with just a simple copy back and
forth. I'm assuming that review comments as well as subsequent
refinement of these scripts over time will gradually converge on the
right set of generic scripts that we can all share, without too many
warts left from whichever facility wrote and contributed each script
first.

Signed-off-by: Larry Gritz <lg@imageworks.com>