Fix tests not resolving the right binary package (#73)
Since the change to the solver which made resolve options less
opinionated, they do not directly influence the variant being selected
if there is a better solution. This change creates strict solver requests
based on the selected variant being tested which forces the solver
to have very little variance on which binary package is being tested

Also add a simple example to prove the command line interface as the logic
lives there. Something to look at refactoring in the future

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>