Package improvements: boost, openimageio, opencolorio, openexr, imath (#95)
* Upgrade openexr and imath specs to v3.0.4.
* Add boost and boost-python packages.
* Add openimageio package (this one was a doozy, lots of dependencies,
  now you can see why I'd been staging so many oddball packages).
* Upgrade opencolorio to 2.0.1.
* Change opencolorio to build just the libraries and python bindings,
  and add a new package opencolorio-apps that has the command line
  utils. This allows us to disentangle an unfortunate dependency cycle
  between OIIO and OCIO. These packages need to be built in the
  following order: opencolorio, openimageio (depends on opencolorio),
  opencolorio-apps (depends on both of the others).

None of this is perfect. It's a work in progress that will have further
refinement. But it seems to work and this reflects how I'm building these
packages for SPI at the moment.

Note: the openimageio.spk.yaml is for a "developer preview" tag based on
master. If you prefer a supported release (2.2.x), you'll need to change
this. When we remove the version from being hard-coded in the spec, this
will be a little less awkward.

Signed-off-by: Larry Gritz <lg@imageworks.com>