Revert "Move required_compat into property of version range rather than request"
I thought that we could clean up the api this way, but actually it
causes new issues where the context of the required compat cannot be
given. I was thinking that this could be reasoned about based on if the
spec is for a binary or source package, but this does not actually
represent the desired behaviour. It's the dependencies of a source
package that can be API compatible during a build, not the source
package itself (since other binary packages may already be compiled
against some ABI of the source package to be rebuilt)

This reverts commit 76c91a4948d818e18c709ac510e636314bc5caec.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>