Use "Binary" instead of "b" when serializing
Use "{:#}" to request the "alternate" string format for `CompatRule`, so
package yamls get written out in a form that will successfully be parsed
back in.

In order for this to work properly, the objects that sit in the middle like
`VersionFilter` need to propagate the current `Formatter` down to the child
objects and use `format!` instead of `to_string`.

Signed-off-by: J Robert Ray <jrray@imageworks.com>