Cache State.__hash__
Profiling showed a significant amount of time spent in calculating hashes,
with both Node.__hash__ and State.id computing the hash.

In my test case, this cache is getting a hit rate of 246,373 and a miss
rate of 9,645.

|                | accum time of State.__hash__ | total time |
|----------------|------------------------------|------------|
|    With cache: |  1.125 s                     | 12.840 s   |
| Without cache: | 15.702 s                     | 27.623 s   |

Changes: Now using a (mutable) field inside State because the
possibility of reusing memory addresses can cause false positives in the
cache.

Signed-off-by: J Robert Ray <jrray@imageworks.com>