Cache State methods that return non-mutated values
These methods are believed to return values that are not mutated, and
therefore it is safe to cache the return values and return a cached result.

Note that without caching __hash__, adding @lru_cache to a method with no
args will invoke hash(self) and end up being slow.

   Runtime with this change: 12.025 seconds.
Runtime without this change: 19.693 seconds.

Changes: The test suite shows that two of the original four methods are
not safe to cache.

Signed-off-by: J Robert Ray <jrray@imageworks.com>