Handle encountering a whiteout entry
While `spk convert pip` was not blocking the attempt to install setuptools,
the conversion of setuptools was dying with a cryptic "File not found"
error:

...
Successfully installed setuptools-57.4.0
[info     ] Purging all changes made to source directory dir=/spfs/spk/pkg/python-setuptools/57.4.0/src
[info     ] Validating package fileset...
No such file or directory (os error 2)

The `read_dir` encounters this deleted directory:
/spfs/lib/python3.7/site-packages/setuptools-57.0.0.dist-info

However, `symlink_metadata` fails with NotFound on this path. The
`dir_entry` is able to be inspected to learn if the entry is a character
device. This can be leveraged to detect this case and return it as a Mask
entry.

Signed-off-by: J Robert Ray <jrray@imageworks.com>