Detect newly created Mask entries as deletions when diffing
To follow up on the previous commit, with that change alone,
`spk convert pip --python-version=3.7 setuptools` will actually succeed,
because the removal of
/spfs/lib/python3.7/site-packages/setuptools-57.0.0.dist-info matches
the (None, Some(_)) arm and is treated as "Added" even though it is a
removal.

Add a new arm to match on Mask entries on the rhs show these end up diffing
as a "Removed" event.

Signed-off-by: J Robert Ray <jrray@imageworks.com>