package_iterator trait overhaul checkpoint
Implementing Iterator for BuildIterator as an enum proved to be a challenge
as next() takes a &mut self and cannot borrow any of the enum kinds as
mutable to delegate to their next().

Ultimately the attempt to implement Iterator was abandoned and enums could
possibly still work here, but in between the code was coverted to use
traits. DynClone has been brought in to satisfy the need from PyO3 for
everything to implement Clone.

Signed-off-by: J Robert Ray <jrray@imageworks.com>