Fix State::get_merged_request error case
The python version would raise KeyError for this "INTERNAL ERROR" case,
and some validator code would catch and ignore KeyError. Don't panic for
this case because it is expected to happen in some situations.

Create a custom enum for the return type so that the validator will be able
to detect this situation specifically.

Signed-off-by: J Robert Ray <jrray@imageworks.com>