Implement Solver::run
This is needed to run `spk env`. The "steps" are now captured as the solver
progresses, and `run` returns an object that can be used to iterate over
the steps.

Note that this is saving the full history of all the steps and it uses a
lot of memory, and quickly uses too much memory if the solver has to go
through a lot of states.

Signed-off-by: J Robert Ray <jrray@imageworks.com>