Move original function state up to runtime, refactor
The original python generator used state in the run function to create a
generator, but we can't do this in rust. Instead, we put that state in
the SolverRuntime type which used to hold the generator. The original
generator logic and solution generation logic now live in the runtime as
the next function and current_solution functions, respectively.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>