Update payload storage trait to use boxed reader
When I first wrote these traits the use of a box seemed dirty, and a
reference was the only other way to pass around a dyn object. In any
case, this more concretely solidifies the interface to be a readable
type that is send and static (owned). Which all of the existing
implementations already easily satisfied.

More concretely, this will enable the interop with web frameworks on
both the client and server side.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>