Satisfy the blob upload workflow on both the client and server
Although it's not super convenient to return an encoded grpc data type
from an http server, this seemed like a better approach than introducing
serde as a dependency just to handle the json serialization of this
single return type.

This also required the addition of Sync to all reader/writers used
in payload storage (because of the requirement for reqwest::Body).

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>