Remove mutable functions from storage interfaces
Having these baked in create an arbitrary need for locks
and management of mutability for API users, even when the
underlying implementation doesn't require it. This way
the implementation is required to manage that as needed.
Our primary filesyste implementation doesn't need it
so there's already a large benefit there.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>