Update logging setup from the rust side
Up until now, it was assued that the only rust logging was spfs, but now
as the codebase grows there is increasingly more information being
logged in rust, and it was not showing up in tests or at the command
line. Additionally, this changes the way that it is configured to only
use the env filter, as the interplay between the two was convoluted at
best. Now, we simply prepend the RUST_LOG env var with the default
directives for the current verbosity level, allowing everything to be
easily overriden with this variable as needed

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>