Add EntryType for tags
When getting a list of tags, it is useful to be able to tell the difference
between entries that are actually tags and entries that are folder names.
Use an enum to discriminate between these two things at the type level.

The string value inside `EntryType::Folder` does not have a trailing `/`
but one is appended to it in its `Display` implementation.

With this change, the ordering of tags is different. Now entries that are
folders are ordered smaller than entries that are tags.

Signed-off-by: J Robert Ray <jrray@imageworks.com>