Remove Clone from State
For performance and memory efficiency reasons, given that State can be a
large and complex object, remove Clone as a derived trait. Instead of
cloning a State object, which is immutable, share references to it through
an Arc.

PyO3 requires Clone, so create a wrapper struct for python that contains
an Arc<State>.

Signed-off-by: J Robert Ray <jrray@imageworks.com>