Add --variant option to build and make-binary
This enables building only a single variant as described by the package
being built, whereas normally all variants would get built. The number
of variants can be determined with the num-variants command. Together,
these can be used to build an automated pipeline in CI to build variants
in parallel.

Specifying a variant index that is out of range will simply be a no-op.
Use of this feature while building multiple packages is not an error but
likely also makes no sense.

Signed-off-by: J Robert Ray <jrray@imageworks.com>