`build` uses the source package it created
Working towards solving #134, extend `struct MakeBinary` so that the
specified packages can also name the source package ident expected to be
used when building. This is not actually guaranteed yet with this change!

When `build` is invoking `make-source` it now gets the list of source
package identities and passes them along to `make-binary`.

Signed-off-by: J Robert Ray <jrray@imageworks.com>