Make repositories aware of their names
A frequent request is to show what repo packages are resolved from, to at
least have an easy way to confirm that a package was picked up from the
local repo as expected. These names like "local" or "origin" are
different from the repository address and need to be captured after parsing
the config file.

Signed-off-by: J Robert Ray <jrray@imageworks.com>