Track the repos that builds are found in
Working towards being able to ignore builds that are not found in a repo
that has been specifically requested for one request.

This may also fix a subtle bug where some builds can be ignored if there
are different sets of builds for the same version in different repos. All
builds of all repos where a version is found need to be collected, and
then these can be de-duplicated.

Signed-off-by: J Robert Ray <jrray@imageworks.com>