Alternative forms of the proposed spec
These are an attempt to move the "when" clause up into a prefix position
so its placement is more consistent across the different forms where it
can appear.

These aren't particuarly compelling examples. There are cases where the
list of package dependencies (names and or quantity) are different
depending on the version of python and that would be an example where
being able to group multiple things under the same "when" clause is
advantageous.

Signed-off-by: J Robert Ray <jrray@imageworks.com>