Add RemoteSpecifier enum
Config::get_remote used to take a &str-like and would first look for a
configured repository with that name, otherwise treat it as address. The
led to code like `config.get_remote("origin")` behaving unexpectedly when
running in a context where there was no "origin" repo configured. Instead,
it would try to open a repo at the file path "/origin". Until recently, it
would also auto-create this repo.

The caller of `Config::get_remote` should know if it is trying to open a
configured repo by name, or if it is trying to pass in an address to some
other repo not in the configuration file. The RemoteSpecifier enum was
added to make this distinction. It ended up that there were no cases where
RemoteSpecifier::Address was needed in the existing code, but it is needed
to retain the original functionality.

The main difference now is that an attempt to open a repo called "origin"
that is not found in the configuration will return an appropriate "not
found" error.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>