Expand the Package trait really naively for validators
This is simpy doing the most basic expansion of the trait type and
replacing the raw spec reference for a trait reference in the
solve/validation implementations. At this point I've put no thought into
whether this is a generic or useful trait shape but the intial goal is
to remove references to the spec type outside of the api module.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>