Remove the spfs-init command
This was leftover legacy from spfs being written in python and so
needing to separate the process so that the enter stuff could be
in C and handle the namespacing and privileged operations. Now that
things are in rust, there's no need to keep the commands separate
and need to load the same information twice for each one.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>