Reduce i/o by exposing control over runtime saving
Previously every change was synced to disk. In reality, these changes
likely did not improve any kind of race conditions and in some cases
actually made it so that the runtime could get out of sync because the
metadata was changed before for the actual operation succeeded. This
also meant that a lot of runtime management code was making a bunch of
saves to the metadata in rapid succession rather than saving once at the
end.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>