Introduce out-of-band monitor process
This separate process is responsible for monitoring the processes that
are operating in a runtime/namespace and then clean up the runtime
metadata when there are no processes left running. Having this process
operate in the background helps to keep the cleanup from slowing
process termination for the original spfs command.

The current implementation scans the /proc filesystem, which is not
particularly efficent but gets the job done.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>