Adds support for tracking what a request was 'requested by'
This adds the 'RequestedBy' enum for tracking what made a PkgRequest.
It adds 'requested_by' data to PkgRequests and updates the
constructor, the 'from_ident()' method, and creation points to support
it. The From trait implementation for PkgRequests is removed in favour
of the two parameter 'new()' and 'from_ident()' methods. BTreeSet is
used to replace HashSet in VersionFilters to provide consistent
ordering for matching up with 'requested_by' data in output messages.
It adds display of 'requested by/required by' info in solver and
solution output. And adds 'problem_package' frequency tracking to
solver stats.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>