Make `compute_tree_node` work in parallel
When `ManifestBuilder` is used by `commit_dir`, a SHA256 hash is calculated
for each file, so it is beneficial to spawn tasks so these computations can
be ran on different cores.

`tokio::spawn` requires 'static lifetimes which prompted a few design
changes. `commit_dir` is moved out of `trait Repository` in order to have a
way to move a `RepositoryHandle` into the spawned task. A trait
`ManifestBuilderHasher` was added to supply hashing behavior instead of
passing in a closure to `ManifestBuilder`.

This is a 160% improvement in throughput on top of the `sync_all` removal,
and a 2,659% improvement compared to the master branch.

jrray@STRANDED:~/git/spfs$ cargo bench --bench spfs_bench -- --load-baseline new --baseline no-sync-all
    Finished bench [optimized] target(s) in 0.13s
     Running benches/spfs_bench.rs (target/release/deps/spfs_bench-eb3cde87f96aad68)
spfs commit path/repo.commit_dir
                        time:   [296.51 ms 300.07 ms 303.08 ms]
                        thrpt:  [3.3786 Kelem/s 3.4125 Kelem/s 3.4535 Kelem/s]
                 change:
                        time:   [-63.637% -61.646% -59.329%] (p = 0.00 < 0.10)
                        thrpt:  [+145.88% +160.73% +175.01%]
                        Performance has improved.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>