Allow room for extra characters in overlay args
The mount args may get an `"ro,"` prefix included and this could make the
difference between the args being too long or not. When
`resolve_overlay_dirs` calls `get_overlay_args`, it needs to also know
how many extra characters are intended to be added to the overlay args.
This was a little awkward because the decision to add the `"ro,"` prefix
lives elsewhere in `env::mount_env`. This decision has been lifted out so
the prefix is now known in advance.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>