Flatten layers in groups
When there are too many layers, the number of layers are halved in order
to find a size that fits in `O(log n)` steps. However, merging an arbitrary
number of layers together is likely to produce a new unique manifest that
won't already exist. By grouping with a fixed number, it becomes more
likely for layers at the bottom of the stack to get grouped into something
shared by other similar environments and reuse existing renders.

The number 7 was selected because it feel like it might have a good chance
to strike a balance between the number of merges and the chance of reuse.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>