Deduplicate all sync tasks by digest
Previously, if the same object appeared twice, we could end up creating
two parallel tasks to sync it at the same time. This is often not a huge
problem but if you got unlucky and this happened for a large payload it
can be very wasteful. It was also more likely to happen for a large
payload because the check for existance would be negative until the
entire payload had been copied.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>