Use more accurate total for progress bars
Previously the progress bars showed the entire amount of data to be
synced, including existing items which meant that it often finished part
way though. This change does mean that large syncs may not immediately
show the total, but it only ever shows the total calculated number of
items/data that require syncing.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>