Add explicit runtime shutdown with timeout
As per the tokio docs, it's possible for a runtime not to shut down if
there is a blocking task or long running task that fails to terminate
when asked to. We don't want to have this happen at the command line
under any circumstances, as all primary logic is expected to have
already been await-ed so a blocked runtime shutdown is assumed to be
some misbehaving cleanup or secondary monitor logic.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>