Restructure `version_range` parser
To make this parser more generally useful, it itself needs to handle
parsing a comma-separated list of `VersionRange` (and turn that into
a `VersionRange::Filter` if there are more than one).

The behavior modifiers used in `wildcard_range` need to be lifted up so
that these can be different for different users of `version_range`.

Signed-off-by: J Robert Ray <jrray@imageworks.com>