Break single package trait into series of traits
The Template is what's loaded from disk and can be rendered into a
Recipe (for existing spec type this is a no-op). The Recipe represents
all possible builds of a package, and can generate both binary and
source Package instances to represent builds. For the current spec type
Recipes and Packages are the same Spec, but the original
update_for_build logic is what we capture in these new types.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>