Remove `LowestSpecifiedRange::new_version_range`
This parsing was broken and calculated the wrong `specified` count for
inputs like `"0.0+A.0"`. Since this string has two `.` characters it
would set `specified` to 3.

In the context this is needed, we can already have a parsed `Version` to
use.

Update build key test cases to pass again after this change, which seems to
correct a bug that the cases were written to expect. However, the order
expected by `test_build_piece_ordering` is still suspect.

Signed-off-by: J Robert Ray <jrray@imageworks.com>