Stop using `api::parse_version`
Add the ability to parse into `Version` objects directly in the parser
module. `api::parse_version` has a flaw if given a string with a pre-
release specified before a post-release, as in: `"0+A.0-a.0"`

`api::parse_version` parses this as: `"0+A.0"` and discards the pre-release
portion.

Update the tagset parsing to accept pre-release and post-release tags in
any order.

Signed-off-by: J Robert Ray <jrray@imageworks.com>