Use `all_consuming` when using `map_parser`
Whatever input the first parser passed to `map_parser` consumes is what
will be consumed, even if the second parser does not consume all the input
passed to it. Use `all_consuming` so that any left over input is not
silently ignored.

Fixes the parser not failing on input like `"my-pkg/1.0,4.0+r."`; it was
parsing this as if the input was `"my-pkg/1.0"` since the invalid part was
being silently dropped.

Signed-off-by: J Robert Ray <jrray@imageworks.com>