Separate Recipe and Package in the Repository trait
The repository trait now associates the relevant recipe and package
types that it will store and retrieve. Additionally, we have separately
named functions for reading a recipe or a built package type. This
obviously has effects on solve process, etc but is ultimately just
formalizing a concept that already existed but was being handled more
obscurely by the single spec type. In the future, it would be good to do
a similar change to the Ident type so that the inputs to these trait
methods could be equally unambiguous.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>