Add `parse_until` combinator
A common pattern emerging is the desire to parse everything up to a
delimiter, and require that everything up to that delimiter is consumed.
Using `all_consuming` for this creates a confusing `Eof` error positioned
earlier in the input. Using `parse_until` helps mark where the unexpected
input starts and shows what character would be expected there.

    Caused by:
        0: at line 1:
        my-pkg:{fred,wilma}x/1.0,4.0+r.3
                           ^
        expected '/', found x

This is only enabled for tests right now to suppress the unused code
warning.

Signed-off-by: J Robert Ray <jrray@imageworks.com>