Ressurect old storage code, and connect to new logic
We do a little inefficient hackiness here to save a representation of
the new runtime in the old format. Obviously only some fields coorelate,
but this will alow previous verisons of spfs to read and join the
runtime as well as allow previous spk versions to read and modify the
runtime (in a lossy way).

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>