Adds an undeprecate command, and updates and fixes bugs in the deprecate, undeprecate, and ls commands
The undeprecate and deprecate commands both use the same underlying change_deprecation_status function, refactored from the deprecate command. Both commands gather all the changes and ask the user for confirmation before taking any action.
This also fixes the bug that prevented deprecate and undeprecate from working on the builds when given just a package version.

The ls command is updated hide deprecated versions and builds by default. The --deprecated flag has been added to ls to show and highlight deprecated versions and builds. This also fixes a sorting bug in ls, when run as 'spk ls packagenameonly', that did not sort the versions in a semantic order.

This also moves some of the testing macros from solve/solver_test.rs into macros.rs in the spk lib crate and puts them behind the 'test-macros' feature so the can be used in spk lib and spk cli tests in future.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>