Ensure that syncer can be safely run as noop
Not technically a no-op, but the idea is not to fail under default
configuration when all of the data already exists or has a
representation in the destination repository. This way commands like
`spfs run` don't need additional logic to check what is and is not
already available before initiating the sync process.

Signed-off-by: Ryan Bottriell <rbottriell@ilm.com>