Rework `version_range::Ranged::intersect`
The existing code was catching some specific cases but letting other
incompatible cases slip through. It was unclear how noodily the logic here
would have to become to handle all the possible cases correctly. This
changeset introduces some new types to help solve the problem in a
systematic way where the individual cases possible are teased apart and
can be reasoned about in isolation.

The biggest change in the approach here is that two `Ranged` are assumed
to not intersect unless they can be proven that they do. All possible
cases are covered.

With these change, the test case failures are reduced to:

    a:!=1.2.0 + b:=1.2.0 == Compatible
    a:>1.0    + b:=1.0   == Compatible
    a:<=1.2   + b:=1.3   == Compatible

These are failing due to bugs outside of the logic added here and need
separate fixes.

Signed-off-by: J Robert Ray <jrray@imageworks.com>