Add "epsilon" concept to `Version`
The `greater_or_equal_to` and `less_than` logic on a number of
`VersionRange` types were flawed because of the lack of a way to express a
version number that is infinitesimally larger than some other version
number.

For example, to express `>1.0`, the `GreaterThanRange::greater_or_equal_to`
implementation was returning the equivalent of `>=1.0` which is wrong for
version 1.0. Now it can properly describe `>1.0` by enabling the
`plus_epsilon` bit.

Similarly, `<=1.0` could not be expressed properly because `less_than` is
an exclusive range, but now `1.0+epsilon` can be used.

`EqualsVersion::less_than` was turning `=1.0` into `<1.1` but this allows
version 1.0.1 to match.

`Version::parts` was pulled out to a separate struct in order to add the
new boolean field and take advantage of auto-deriving `Eq` and `PartialEq`
for it. However the `Ord` logical still fully lives in `Version` and could
be refactored.

This change makes all the previously failing test cases pass, but there
isn't full coverage so there may still be some cases not handled properly.
Notably the `SemverRange` and `WildcardRange` types haven't been updated
yet.

Signed-off-by: J Robert Ray <jrray@imageworks.com>