Fix failing test cases for `VersionRange::contains`
The logic here has to necessarily expand to the point where it is better
to abstract out the "lower" versus "upper" bounds checking code since the
logic is essentially the same. The order of "self" vs "other" changes, so
these terms have been changed to "left" and "right". It's easier to see in
the old code how the two `if` blocks have nearly identical logic.

The `Incompatible` messages have been modified to disambiguate the
different places in the code that they are created, to help debugging. The
wording of "a wider range" is not entirely accurate for some cases, so
the wording was change to something more neutral. For example:

    "[case 2,1] ~1.2.3 represents a wider range than allowed by >1.2.4"

This is not really true, but making it say "narrower" in the right cases
is complicated, and in some cases "disjoint" would be more accurate.

Signed-off-by: J Robert Ray <jrray@imageworks.com>