Merge rules when possible in `VersionFilter::restrict`
Refactor the "simplify" logic into a reusable method, and then when
restricting a version filter with another one, find the smallest set of
rules that accurately represent the union of the rules.

`CompatRange` rules cannot be merged because that requires knowing in
advance what the `compat` rules will be for the builds that are found.

Signed-off-by: J Robert Ray <jrray@imageworks.com>