Handle $SPFS_STORAGE_RUNTIMES in spfs-enter
When legacy spk calls spfs-enter, it might give the runtime storage based
on $SPFS_STORAGE_RUNTIMES, as in:

    spfs-enter --remount /tmp/spfs-runtimes/<UUID>

In this form, the existing `rsplitn(3, ...)` happens to be able to still
split this three times, but the results are not desirable. Detect this
form but don't treat the path as the path to a real spfs repo, since it is
not (necessarily) a real repo at that path, which is how things are when
running the legacy spk test suite.

Signed-off-by: J Robert Ray <jrray@imageworks.com>