Don't fail if the new-style runtime can't be read
When the legacy spk test suite runs, it starts off with a new style runtime
created in the normally configured "local" repo, but while the tests run
they change the "local" repo path to some other temporary repo. Then, in
spfs-enter it can't find the runtime by name anymore.

In the event that the new style runtime can't be loaded, still try to read
the legacy runtime config. If that is successful, then recreate the new
style runtime in whatever the new runtime storage happens to be.

This is the last piece needed to make all the legacy spk tests pass.

Signed-off-by: J Robert Ray <jrray@imageworks.com>