Fix `test_shell_initialization_startup_scripts`
This test would fail on a host that has `tcsh` installed but does not have
`expect` installed. Check `find_best_shell`'s behavior when testing the
`tcsh` case and if it doesn't find the expected shell, abandon the test.
Otherwise it tries to run the `startup_cmd` intended to `tcsh` with `bash`.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>