Experiment with saving memory for `State::pkg_requests`
Signed-off-by: J Robert Ray <jrray@imageworks.com>

Cache the materialized `PkgRequestChainIter`

Although `PkgRequestChain` improved the memory footprint and performance,
walking the chain to create a `Vec` to iterate in the desired order was
still taking a performance toll. Caching this has a significant benefit.

Before: Solver took 140392 steps total, at 4679.671 steps/sec
 After: Solver took 159480 steps total, at 5315.913 steps/sec

Signed-off-by: J Robert Ray <jrray@imageworks.com>

Remove PkgRequestChain

Since iterating the pkg_requests requires materializing a `Vec` anyway, go
back to just using a `Vec`. This time, the elements are `Vec<PkgRequest>`
to continue to avoid the overhead of cloning the elements. There is a
small hit to performance but the code is simpler.

Before: Solver took 159480 steps total, at 5315.913 steps/sec
 After: Solver took 155852 steps total, at 5195.049 steps/sec

Signed-off-by: J Robert Ray <jrray@imageworks.com>