Allow repository to be opened if disk is full
Related to #119. Opening the local repo goes through the `create` flow that
always tries writing the `VERSION` file. However, this file may already
exist and have the expected contents. In the event that the writing the
`VERSION` file fails, attempt to _read_ the file to see if it has the
desired contents already.

This required a semantic change to `read_last_migration_version` so it
doesn't misrepresent the true nature of the `VERSION` file. It returns
`Option<Version>` now so that the caller can know if the file really
existed and wasn't empty.

Signed-off-by: J Robert Ray <jrobertray@gmail.com>