Make the write_data function unsafe
This function, when not paired with a subsequent call to create_blob,
ends up creating a payload which is not connected to the graph. Even if
the digest is tagged the clean process will not consider it connected,
only that the tag is broken and pointing to something that doesn't
exist.

Signed-off-by: Ryan Bottriell <ryan@bottriell.ca>