Two solver heuristic changes  (#334) and undeprecate command (#207)
Adds two solver heuristic changes to improve solve performance and an undeprecate command with bugfixes for the deprecate and ls commands and in-memory repo publishing.

The two solver heuristic changes are:
1. Build option value based build ordering for SortedBuildIterator, along with BuildKey for compound build keys, to improve build selection heuristic. This replaces the distance based build ordering.
2. Same package request reordering to RequestPackage to improve next request package selection heuristic. This includes adding duplicate request and same package request counters to solver stats and display.

The undeprecate and deprecate commands both use the same underlying change_deprecation_status function, refactored from the deprecate command. Both commands gather all the changes and ask the user for confirmation before taking any action. This also fixes the bug that prevented deprecate and undeprecate from working on the builds when given just a package version.

The 'ls' command is updated hide deprecated versions and builds by default. The '--deprecated' flag has been added to 'ls' to show and highlight deprecated versions and builds. This also fixes a sorting bug in 'ls', when run as 'spk ls packagenameonly', that did not sort the versions in a semantic order.

The in-memory repo is updated to support updating build specs, and lines its publish_spec and force_publish_spec methods up with the spfs repo ones.

This also moves some testing macros from solve/solver_test.rs into macros.rs in the spk lib crate and puts them behind the 'test-macros' feature so the can be used in spk lib and spk cli tests. Tests for the deprecate and undeprecate commands have been added and they use these macros.

Signed-off-by: David Gilligan-Cook <dcook@imageworks.com>